Add banner manager and other features

+++ src/main/java/ru/soliev/practice/banners/dto/BannerShortDTO.java
@@ -1,4 +1,6 @@
 package ru.soliev.practice.banners.dto;
 
-public record BannerShortDTO(int price, String name, String content) {
+import java.math.BigDecimal;
+
+public record BannerShortDTO(BigDecimal price, String name, String content) {
 }
